+++ includes/class.Exporter.php
@@ -1,19 +1,15 @@
 <?php
+namespace webaware\em_import_export;
+
+use DateTimeZone;
+use EM_Events;
+use XMLWriter;
 
 if (!defined('ABSPATH')) {
 	exit;
 }
 
-class EM_ImpExpExport {
-
-	protected $plugin;
-
-	/**
-	* @param EM_ImpExpPlugin $plugin handle to the plugin object
-	*/
-	public function __construct($plugin) {
-		$this->plugin = $plugin;
-	}
+class Exporter {
 
 	/**
 	* render the admin page
@@ -26,7 +22,7 @@ class EM_ImpExpExport {
 	* export the data in selected format
 	*/
 	public function export() {
-		$EM_Events = EM_Events::get();
+		$EM_Events = EM_Events::get(array('scope' => 'all'));
 
 		$format = isset($_POST['exp_format']) ? wp_unslash($_POST['exp_format']) : '';
 
@@ -62,6 +58,8 @@ class EM_ImpExpExport {
 
 		nocache_headers();
 
+		$UTC = new DateTimeZone('UTC');
+
 		$xml = new XMLWriter();
 		$xml->openURI('php://output');			// write directly to PHP output
 		$xml->startDocument('1.0', 'UTF-8');
@@ -86,7 +84,7 @@ class EM_ImpExpExport {
 			$xml->startElement('vevent');
 
 			// manufacture a unique ID
-			$xml->writeElement('uid', self::getUniqueID($EM_Event));
+			$xml->writeElement('uid', get_em_unique_id($EM_Event));
 
 			// add link to event post
 			$xml->writeElement('url', $EM_Event->output('#_EVENTURL'));
@@ -102,16 +100,18 @@ class EM_ImpExpExport {
 			$xml->writeElement('x-event_spaces', $EM_Event->event_spaces);
 
 			// get the start, end and last modified dates/times
-			// TODO: verify that gmt_offset works when switching between real time and Daylight Stupid Time
-			$gmtOffset = HOUR_IN_SECONDS * get_option('gmt_offset');
-			$xml->writeElement('dtstart', date('Y-m-d\TH:i:s\Z', $EM_Event->start - $gmtOffset));
-			$xml->writeElement('dtend', date('Y-m-d\TH:i:s\Z', $EM_Event->end - $gmtOffset));
+			$time_start = clone $EM_Event->start();
+			$time_start->setTimezone($UTC);
+			$xml->writeElement('dtstart', $time_start->format('Y-m-d\TH:i:s\Z'));
+			$time_end = clone $EM_Event->end();
+			$time_end->setTimezone($UTC);
+			$xml->writeElement('dtend', $time_end->format('Y-m-d\TH:i:s\Z'));
 			$xml->writeElement('dtstamp', date('Y-m-d\TH:i:s\Z', strtotime($EM_Event->post_modified_gmt)));
 
 			// get categories as comma-separated list
 			$cats = $EM_Event->get_categories()->categories;
 			if (count($cats) > 0) {
-				$categories = array();
+				$categories = [];
 				foreach ($cats as $cat) {
 					$categories[] = $cat->output('#_CATEGORYNAME');
 				}
@@ -185,7 +185,7 @@ class EM_ImpExpExport {
 					$xml->startElement('rrule');
 					$xml->startElement('recur');
 
-					$days = array('SU','MO','TU','WE','TH','FR','SA');
+					$days = ['SU','MO','TU','WE','TH','FR','SA'];
 					$until = date('Y-m-d\TH:i:s\Z', $recurrence->end - $gmtOffset);
 
 					switch ($recurrence->freq) {
@@ -197,7 +197,7 @@ class EM_ImpExpExport {
 
 						case 'weekly':
 							$bydays = explode(',', $recurrence->byday);
-							$BYDAY = array();
+							$BYDAY = [];
 							foreach ($bydays as $day) {
 								$BYDAY[] = $days[$day];
 							}
@@ -248,8 +248,8 @@ class EM_ImpExpExport {
 		nocache_headers();
 
 		// character conversion arrays
-		$charFrom = array('\\', ';', ',', "\n", "\t");
-		$charTo = array('\\\\', '\;', '\,', '\\n', '\\t');
+		$charFrom = ['\\', ';', ',', "\n", "\t"];
+		$charTo = ['\\\\', '\;', '\,', '\\n', '\\t'];
 
 		// output header
 		echo "BEGIN:VCALENDAR\n";
@@ -262,11 +262,11 @@ class EM_ImpExpExport {
 		foreach ($EM_Events as $EM_Event) {
 
 			// build array of iCalendar lines, and start event
-			$ics = array();
+			$ics = [];
 			echo "BEGIN:VEVENT\n";
 
 			// manufacture a unique ID
-			$ics[] = 'UID:' . self::getUniqueID($EM_Event);
+			$ics[] = 'UID:' . get_em_unique_id($EM_Event);
 
 			// add link to event post
 			$ics[] = 'URL:' . $EM_Event->output('#_EVENTURL');
@@ -279,13 +279,13 @@ class EM_ImpExpExport {
 			if (is_object($location)) {
 				if (empty($location->location_name)) {
 
-					$parts = array (
+					$parts = [
 						$location->location_address,
 						$location->location_town,
 						$location->location_state,
 						$location->location_postcode,
 						$location->get_country(),
-					);
+					];
 					$address = implode(', ', array_filter($parts, 'strlen'));
 
 					$ics[] = 'LOCATION:' . str_replace($charFrom, $charTo, $address);
@@ -305,7 +305,7 @@ class EM_ImpExpExport {
 			// get categories as comma-separated list
 			$cats = $EM_Event->get_categories()->categories;
 			if (count($cats) > 0) {
-				$categories = array();
+				$categories = [];
 				foreach ($cats as $cat) {
 					$categories[] = str_replace($charFrom, $charTo, $cat->output('#_CATEGORYNAME'));
 				}
@@ -317,7 +317,7 @@ class EM_ImpExpExport {
 			if (!$EM_Event->is_individual()) {
 				$recurrence = $EM_Event->get_event_recurrence();
 				if (!empty($recurrence)) {
-					$days = array('SU','MO','TU','WE','TH','FR','SA');
+					$days = ['SU','MO','TU','WE','TH','FR','SA'];
 					$until = date('Ymd\THis\Z', $recurrence->end - $gmtOffset);
 
 					switch ($recurrence->freq) {
@@ -327,7 +327,7 @@ class EM_ImpExpExport {
 
 						case 'weekly':
 							$bydays = explode(',', $recurrence->byday);
-							$BYDAY = array();
+							$BYDAY = [];
 							foreach ($bydays as $day) {
 								$BYDAY[] = $days[$day];
 							}
@@ -385,16 +385,16 @@ class EM_ImpExpExport {
 		foreach ($EM_Events as $EM_Event) {
 			echo '1,';
 
-			echo self::text2csv($EM_Event->event_name), ',';							// event_name
-			echo self::text2csv(preg_replace('/\s+/', ' ', strip_tags($EM_Event->post_content))), ',';				// event_desc
+			echo text_to_csv($EM_Event->event_name), ',';							// event_name
+			echo text_to_csv(preg_replace('/\s+/', ' ', strip_tags($EM_Event->post_content))), ',';				// event_desc
 
 			$location = $EM_Event->get_location();
 			if (is_object($location)) {
-				echo self::text2csv($location->location_address), ',';					// address
-				echo self::text2csv($location->location_town), ',';						// city
-				echo self::text2csv($location->location_state), ',';					// state
-				echo self::text2csv($location->get_country()), ',';						// country
-				echo self::text2csv($location->location_postcode), ',';					// zip
+				echo text_to_csv($location->location_address), ',';					// address
+				echo text_to_csv($location->location_town), ',';						// city
+				echo text_to_csv($location->location_state), ',';					// state
+				echo text_to_csv($location->get_country()), ',';						// country
+				echo text_to_csv($location->location_postcode), ',';					// zip
 			}
 			else {
 				echo ',,,,,';
@@ -403,13 +403,13 @@ class EM_ImpExpExport {
 			echo ',';																	// phone
 
 			echo 'Y,';																	// display_desc
-			echo self::text2csv(self::getUniqueID($EM_Event)), ',';						// event_identifier
+			echo text_to_csv(get_em_unique_id($EM_Event)), ',';						// event_identifier
 
 			$gmtOffset = HOUR_IN_SECONDS * get_option('gmt_offset');
-			echo self::text2csv(date('Y-m-d', $EM_Event->start - $gmtOffset)), ',';		// start_date
-			echo self::text2csv(date('Y-m-d', $EM_Event->end - $gmtOffset)), ',';		// end_date
-			echo self::text2csv(date('H:i:s', $EM_Event->start - $gmtOffset)), ',';		// start_time
-			echo self::text2csv(date('H:i:s', $EM_Event->end - $gmtOffset)), ',';		// end_time
+			echo text_to_csv(date('Y-m-d', $EM_Event->start - $gmtOffset)), ',';		// start_date
+			echo text_to_csv(date('Y-m-d', $EM_Event->end - $gmtOffset)), ',';		// end_date
+			echo text_to_csv(date('H:i:s', $EM_Event->start - $gmtOffset)), ',';		// start_time
+			echo text_to_csv(date('H:i:s', $EM_Event->end - $gmtOffset)), ',';		// end_time
 
 			echo '100,';	// reg_limit
 			echo '0,';	// event_cost
@@ -418,34 +418,76 @@ class EM_ImpExpExport {
 			echo 'N,';	// send_mail
 			echo 'Y,';	// is_active
 			echo ',';	// conf_mail
-			echo self::text2csv(date('Y-m-d', $EM_Event->start - $gmtOffset)), ',';	// registration_start
-			echo self::text2csv($EM_Event->event_rsvp_date), "\r\n";	// registration_end
+			echo text_to_csv(date('Y-m-d', $EM_Event->start - $gmtOffset)), ',';	// registration_start
+			echo text_to_csv($EM_Event->event_rsvp_date), "\r\n";	// registration_end
 		}
 
 		exit;
 	}
+	
+        /**
+        * export data in Event CSV format
+        * @param EM_Events $EM_Events
+        */
+        public function exportCSV($EM_Events) {
+                if (isset($_REQUEST['plaintext']) && $_REQUEST['plaintext'] === '1') {
+                        header('Content-Type: text/plain; charset=utf-8');
+                }
+                else {
+                        header('Content-Type: text/csv; charset=utf-8');
+                        header('Content-Disposition: attachment; filename="events.csv"');
+                }
 
-	/**
-	* encapsulate text in quotes if unsuitable for CSV without quotes
-	* NB: Events Espresso is highly dodgy and doesn't handle apostrophes, so must convert!
-	* @param string $text
-	* @return string
-	*/
-	protected static function text2csv($text) {
-		$len = strlen($text);
-		if ($len > 0 && $len != strcspn($text, "\"',;$\\\r\n0123456789"))
-			return '"' . strtr(str_replace('"', '""', $text), "'", '`') . '"';
+                nocache_headers();
 
-		return $text;
-	}
+                // character conversion arrays
+                $charFrom = ['\\', ';', ',', "\n", "\t"];
+                $charTo = ['\\\\', '\;', '\,', '\\n', '\\t'];
+                // send header row
 
-	/**
-	* generate unique ID for event
-	* @param EM_Event $EM_Event
-	* @return string
-	*/
-	protected static function getUniqueID($EM_Event) {
-		return "events-manager-{$EM_Event->event_id}@" . parse_url(get_option('home'), PHP_URL_HOST);
-	}
+                echo "uid,summary,dtstart,dtend,dtformat,categories,post_content,location_name,location_address,location_town,location_state,location_postcode,location_country,location_latitude,location_longitude\r\n";
 
+                foreach ($EM_Events as $EM_Event) {
+                        echo $EM_Event->event_id, ',';                         // event_identifier
+
+                        echo text_to_csv($EM_Event->event_name), ',';          // event_summary
+                        $gmtOffset = HOUR_IN_SECONDS * get_option('gmt_offset');
+                        echo text_to_csv(date('Y-m-d', $EM_Event->start - $gmtOffset)), ',';            // start_date
+                        echo text_to_csv(date('Y-m-d', $EM_Event->end - $gmtOffset)), ',';              // end_date
+                        echo text_to_csv("Y-m-d"), ',';         // dtformat
+                        // get categories as comma-separated list
+                        $cats = $EM_Event->get_categories()->categories;
+                        if (count($cats) > 0) {
+                                $categories = [];
+                                foreach ($cats as $cat) {
+                                        $categories[] = str_replace($charFrom, $charTo, $cat->output('#_CATEGORYNAME'));
+                                }
+                                echo text_to_csv(implode(',', $categories)), ',';
+                        }
+                        else {
+                                echo ',';
+                    }
+                        echo text_to_csv(preg_replace('/\s+/', ' ', strip_tags($EM_Event->post_content))), ',';                         // event_desc
+
+                        $location = $EM_Event->get_location();
+                        if (is_object($location)) {
+                                echo text_to_csv($location->location_name), ',';// address
+                                echo text_to_csv($location->location_address), ',';                                     // address
+                                echo text_to_csv($location->location_town), ',';// city
+                                echo text_to_csv($location->location_state), ',';                                       // state
+                                echo text_to_csv($location->location_postcode), ',';                                    // zip
+                                echo text_to_csv($location->get_country()), ',';// country
+                                echo text_to_csv($location->latitude), ',';    // country
+                                echo text_to_csv($location->longitude), ',';   // country
+                        }
+                        else {
+                                echo ',,,,,,,';
+                        }
+
+                        echo "\r\n";
+
+                }
+
+                exit;
+        }
 }